Conflicts:
	Life/Source/mainscene.cpp

+++ Life/Source/maploader.cpp
@@ -85,7 +85,7 @@ bool MapLoader::loadMap(std::string file)
 		std::vector<std::string>().swap(Column);
 	}
 
-	worldSize = std::stoi(map_data[0][0]);
+	worldSize = static_cast<float>(std::stoi(map_data[0][0]));
 
 	map_height = map_data.size();
 	return true;
